IPA: do not fail if view name lookup failed on older versions
Depending on the version 389ds return a different error code if the
search for the view name failed because our dereference attribute
ipaAssignedIDView is not known. Newer version return
LDAP_UNAVAILABLE_CRITICAL_EXTENSION(12) which is translated to
EOPNOTSUPP and older versions return LDAP_PROTOCOL_ERROR(2) which is
returned as EIO. In both cases we have to assume that the server is not
view aware and keep the view name unset.

Resolves https://fedorahosted.org/sssd/ticket/2650

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>