Skip enumeration requests in IPA and AD providers as well
Checking the enum request in the underlying LDAP provider to skip it
might be too late as the richer IPA or AD providers depend on having a
useful result when the sdap request finishes.

Move the enumeration check earlier instead and allow directly in the IPA
or AD handler.

Related:
https://fedorahosted.org/sssd/ticket/2659

Reviewed-by: Sumit Bose <sbose@redhat.com>