UTIL/SYSDB: Move new_subdomain() to sysdb_subdomains.c and make it private
In order to make updating the subdomain list a two-step process.
Therefore we need to make sure that update_subdomains() is the only
interface towards the SSSD that changes the subdomain list.

Move the new_subdomain() function to sysdb_subdomains.c and only make it
available through a private header so it's usable by unit tests.

Reviewed-by: Sumit Bose <sbose@redhat.com>