AD: Use ad_site also when site search fails
https://fedorahosted.org/sssd/ticket/2725

Some deployments use the ad_site option for cases where the AD clients
are not able to find a site for one reason or another. With our current
code, the ad_site option value can only override a site that the client
found, not supply the value for cases no site could be found.

This patch fixes the issue.

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>