AD: Only ignore errors from SDAP lookups if there's another connection to fallback to
Required for:
https://fedorahosted.org/sssd/ticket/2637

The AD lookup code honors the ignore_mark_offline flag in the sense that
if it's set, the sdap return code is not reported to the upper layer,
but EOK is returned as request status and the sdap return code is
returned separately.

This patch modifies the behaviour further to only apply if there is
another connection to fall back to.

Reviewed-by: Pavel Březina <pbrezina@redhat.com>