AD: Do not mark the whole back end as offline if subdomain lookup fails
Required for:
https://fedorahosted.org/sssd/ticket/2637

Rather mark the domain as inactive. It will be marked as active later,
in the meantime the main domain can continue to work online and
subdomain requests will be answered from cache.

The lookup request itself just returns a special error code and lets the
caller handle the error code as appropriate (normally by disabling the
subdomain temporarily).

Reviewed-by: Pavel Březina <pbrezina@redhat.com>