BUILD: Avoid symlinks with python modules
We need to use different names for python{2,3} modules if we want to build
them in the same time with automake (prefix _py2 and _py3). But resulting name
need to correspond with name of module because it is used in C import function.
We used symbolic links for that purpose but it breaks debian python tools
which rename the real modules making symbolic links to point nowhere

Resolves:
https://fedorahosted.org/sssd/ticket/2814

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>