LDAP: Inform about small range size
When a returned RID has a higher value than the ldap_idmap_range_size,
it means that the administrator did not plan appropriately for the size
of their network. We need to alert the admin at a severe notification
level that their configuration will fail on entries with a high RID and
point them at the explanation in the manual.

Reviewed-by: Lukáš Slebodník <lslebodn@redhat.com>
(cherry picked from commit a47102e74050d8ab14a9ea835ab2640c9aa65856)