SDAP: change type of attrsonly in sdap_get_generic_ext_state
'attrsonly' parameter is directly passed to ldap_search_ext() and is
describe as:

The attrsonly parameter should be set to a non-zero value if only
attribute descriptions are wanted. It should be set to zero (0) if both
attributes descriptions and attribute values are wanted.

Boolean type should be fine for the 'attrsonly' parameter especially
since the actual parameter was already set to false in function calls.

Reviewed-by: Sumit Bose <sbose@redhat.com>