tests: Regression test with wrong LC_ALL
Ticket:
https://fedorahosted.org/sssd/ticket/2785

Test local domain tool with wrong LC_ALL
environment variable value.

NOTE: The memory cache files are not deleted
properly in the test teardown to work around the
problem described in ticket
https://fedorahosted.org/sssd/ticket/2726

Once the ticket above is solved, the teardown
will be updated to remove the memory cache
files.

Reviewed-by: Michal Židek <mzidek@redhat.com>