SUDO: move offline check to handler
We let sdap_id_op decide if we are offline or not here but we
should not get to this code since ptask is disabled and we will
not get through sudo handler if offline.

This simplyfies the code and make it more similar to other providers.

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>
Reviewed-by: Lukáš Slebodník <lslebodn@redhat.com>
(cherry picked from commit 81f135f9e83031c4a021a3d19009b2bc179c8468)