LDAP: Change the default rfc2307 autofs attribute mappings
Resolves:
    https://fedorahosted.org/sssd/ticket/2858

The default attribute mappings we used to have:
    ldap_autofs_map_object_class        automountMap
    ldap_autofs_map_name                ou
    ldap_autofs_entry_object_class      automount
    ldap_autofs_entry_key               cn
    ldap_autofs_entry_value             automountInformation

Was wrong. Instead, this patch switches to:
    ldap_autofs_map_object_class        nisMap
    ldap_autofs_map_name                nisMapName
    ldap_autofs_entry_object_class      nisObject
    ldap_autofs_entry_key               cn
    ldap_autofs_entry_value             nisMapEntry

Which are attributes that are available with servers running the default
rfc2307 schema. In addition, this patch adds a syslog and DEBUG message
that warns administrators to double-check their configuration.

We don't warn when the autofs provider is set to AD, because that one
is already correct.

Reviewed-by: Lukáš Slebodník <lslebodn@redhat.com>