IPA: fix override with the same name
If the user name of a AD user is overridden with the name itself in an
IPA override object SSSD adds this name twice to the alias list causing
an ldb error when trying to write the user object to the cache. As a
result the user is not available.

This patch makes sure that there are no duplicated alias names.

Resolves https://fedorahosted.org/sssd/ticket/2874

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>