TOOLS: Add comments on functions in colondb
The colondb API provides three function:
* sss_colondb_open()
* sss_colondb_write_field()
* sss_colondb_read_field()

It is not obvious that sss_colondb_open() add destructor on talloc
context which close the colondb during free context. And there is
expectation that SSS_COLONDB_SENTINEL is type of last item in line.

So this patch adds simple lightening comments in doxygen style.

Resolves:
https://fedorahosted.org/sssd/ticket/2764

Reviewed-by: Lukáš Slebodník <lslebodn@redhat.com>