UTIL: Add a PROBE macro into probes.h
The macros are inspired by very similar macros in libvirt code.

Adds a macro PROBE that can be used by SSSD developers to add systemtap
marks to code. These marks, when coupled with a location in a binary can
be in turn used to call probes. The mark can be called like this:
    PROBE(PROBE_NAME, arguments)
This is cleaner than using the SSSD_$(PROBE_NAME) directly as it
directly shows that a probe is being called at that place.

If the systemtap tracing is disabled, they would expand to an empty macro. If
the systemtap tracing is enabled, the systemtap probe will be called.

The overhead of calling the probes is close to zero. As one of the
systemtap developers explained to me:
    """
    STAP_PROBE() macros cost apprx. one nop in the executable, so apprx.
    no cost at all. The more the merrier. Only when activated by a
    stap script do we generally think of it like a microsecond of time.
    """

The probe arguments can be used in the probes to be printed or passed
on to functions. There was an issue in case a string argument was NULL.
This commit adds a helper macro to deal with NULL-strings as if they were
empty ("").

This file would be included by any source file that wants to call the
PROBE() macro.

Reviewed-by: Lukáš Slebodník <lslebodn@redhat.com>