SPEC: Change package ownership of %{pubconfpath}/krb5.include.d
krb5 domain mapping files are stored to the directory
%{pubconfpath}/krb5.include.d. It can be stored by ipa or ad provider.
However this directory was owned by sub-package sssd-ipa. And ad provider
can be installed without this package. Therefore %{pubconfpath}/krb5.include.d
should be owned by common dependency.

The owner of this directory was also fixed to sssd.
It's already done by make install. It was changed only in spec file.

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>