subdomains: inherit ldap_krb5_keytab
If a non-default keytab is configured for the parent domain the
subdomains will still use the default keytab because the alternative
keytab is not inherited. As a consequence SSSD might not be able to
connect to services in the subdomain because the default keytab is
either not present or does not have suitable keys.

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>