IPA: lookup idview name even if there is no master domain record
Currently the IPA subdomain provider returns with a error if there is no
master domain record found. Since this record contains data which is
only needed to create a trust with AD, like e.g. the IPA domain SID,
this record is only created by ipa-adtrust-install. But the idview name
is read after the master domain record. To make the idview feature work
with a plain FreeIPA setup without running ipa-adtrust-install the
missing master domain record should be handled gracefully and the
following lookup should run as well.

Resolves https://fedorahosted.org/sssd/ticket/2960

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>