TESTS: Fix sysdb tests to work with the new format
The sysdb tests now user the qualified name to store users and groups.
To avoid the sysdb interface being tied too tightly to our specific
format, all names are constructed using a function, not hardcoded. Just
swapping the functions that create or parse the names for a different
format should not make the test fail.

Reviewed-by: Sumit Bose <sbose@redhat.com>